feature(application-web) : simplify code architecture
