+++ src/lib/application-web/constantes.ts
@@ -1,0 +1,9 @@
+export enum DefaultOption {
+  AUTHOR = '',
+  DESCRIPTION = '',
+  LANGUAGE = 'ts',
+  VERSION = '0.0.1',
+  PATH_NAME = 'src',
+  LIB_PATH = 'libs',
+  PACKAGE_MANAGER = 'npm',
+}
